linked card details and account details async using rabbitMQ

+++ account/src/main/java/com/finaxis/account/repository/AccountRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountRepository extends JpaRepository<Account,Integer> {
     Account findByAccountId(int accountId);
-    Account findByAccountNumber(int accountRef);
+    Account findByAccountNumber(long accountRef);
 }
